Add comments and clean up some issues in the tests
