+++ Sources/OpenTelemetryApi/Context/OpenTelemetryContextProvider.swift
@@ -48,7 +48,7 @@ public struct OpenTelemetryContextProvider {
         contextManager.removeContextValue(forKey: OpenTelemetryContextKeys.baggage, value: baggage)
     }
 
-    /// Sets `span` as the active span for the duration of the given closure. While the span will no longer be active after the closure exits, this method does **not** end the span.
+    /// Sets `span` as the active span for the duration of the given closure. While the span will no longer be active after the closure exits, this method does **not** end the span. Prefer `SpanBuilderBase.withActiveSpan` which handles starting, activating, and ending the span.
     public func withActiveSpan<T>(_ span: SpanBase, _ operation: () throws -> T) rethrows -> T {
         try contextManager.withCurrentContextValue(forKey: .span, value: span, operation)
     }
@@ -58,7 +58,7 @@ public struct OpenTelemetryContextProvider {
     }
 
 #if canImport(_Concurrency)
-    /// Sets `span` as the active span for the duration of the given closure. While the span will no longer be active after the closure exits, this method does **not** end the span.
+    /// Sets `span` as the active span for the duration of the given closure. While the span will no longer be active after the closure exits, this method does **not** end the span. Prefer `SpanBuilderBase.withActiveSpan` which handles starting, activating, and ending the span.
     public func withActiveSpan<T>(_ span: SpanBase, _ operation: () async throws -> T) async rethrows -> T {
         try await contextManager.withCurrentContextValue(forKey: .span, value: span, operation)
     }
