+++ Sources/OpenTelemetrySdk/Common/autoreleasepool.swift
@@ -8,7 +8,7 @@ import ObjectiveC
 #endif
 
 #if !canImport(ObjectiveC)
-/// Mocks out ObjectiveC's `autoreleasepool` function by simply calling the closure directly
+/// Mocks out ObjectiveC's `autoreleasepool` function by simply calling the closure directly on platforms without Objective-C support
 func autoreleasepool<Result>(invoking body: () throws -> Result) rethrows -> Result {
     try body()
 }
