+++ Tests/ExportersTests/DatadogExporter/Persistence/FileWriterTests.swift
@@ -7,7 +7,7 @@
 import XCTest
 
 class FileWriterTests: XCTestCase {
-    private let temporaryDirectory = obtainUniqueTemporaryDirectory()
+    @UniqueTemporaryDirectory private var temporaryDirectory: Directory
 
     override func setUp() {
         super.setUp()
