+++ Tests/ExportersTests/PersistenceExporter/Storage/OrchestratedFileReaderTests.swift
@@ -7,7 +7,7 @@
 import XCTest
 
 class OrchestratedFileReaderTests: XCTestCase {
-    private let temporaryDirectory = obtainUniqueTemporaryDirectory()
+    @UniqueTemporaryDirectory private var temporaryDirectory: Directory
 
     override func setUp() {
         super.setUp()
