+++ Tests/ExportersTests/PersistenceExporter/Storage/OrchestratedFileWriterTests.swift
@@ -7,7 +7,7 @@
 import XCTest
 
 class OrchestratedFileWriterTests: XCTestCase {
-    private let temporaryDirectory = obtainUniqueTemporaryDirectory()
+    @UniqueTemporaryDirectory private var temporaryDirectory: Directory
 
     override func setUp() {
         super.setUp()
